refactor:some api
