+++ src/app/server_thread_pool.rs
@@ -7,9 +7,8 @@ use tokio::{
 use tracing::{error, info};
 
 use crate::{
-    const_config::DEFAULT_SYSTEM_CHANNEL_SIZE,
-    dto_factory::input_dto::{check_handler_result, HandlerResult},
-    handler::HandlerContext,
+    app::common_api::check_handler_result, const_config::DEFAULT_SYSTEM_CHANNEL_SIZE,
+    dto_factory::input_dto::HandlerResult, handler::HandlerContext,
 };
 
 use super::{AsyncFunc, ClientsStructType, TaskBody};
