+++ src/vo_factory/big_buf_reader.rs
@@ -1,4 +1,4 @@
-use bytes::{Bytes, BytesMut};
+use bytes::BytesMut;
 
 /// A struct for reading large buffers.
 pub(crate) struct BigBufReader {
